+++ runscope/resource_runscope_environment_test.go
@@ -27,6 +27,15 @@ func TestAccEnvironment_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"runscope_environment.environmentA", "verify_ssl", "true")),
 			},
+			{
+				Config: fmt.Sprintf(testRunscopeEnvironmentConfigAWithSingularRemoteAgents, teamID, teamID),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckEnvironmentExists("runscope_environment.environmentA"),
+					resource.TestCheckResourceAttr(
+						"runscope_environment.environmentA", "name", "test-environment"),
+					resource.TestCheckResourceAttr(
+						"runscope_environment.environmentA", "verify_ssl", "true")),
+			},
 		},
 	})
 }
@@ -51,6 +60,32 @@ func TestAccEnvironment_emails(t *testing.T) {
 					resource.TestCheckResourceAttr("runscope_environment.environmentA", "emails.0.recipients.#", "2"),
 					resource.TestCheckResourceAttr("runscope_environment.environmentA", "emails.0.recipients.769210288.name", "bob"),
 					resource.TestCheckResourceAttr("runscope_environment.environmentA", "emails.0.recipients.769210288.email", "bob@gmail.com"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccEnvironment_email(t *testing.T) {
+	teamID := os.Getenv("RUNSCOPE_TEAM_ID")
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckEnvironmentDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(testRunscopeEnvironmentConfigWithEmailSingular, teamID, teamID),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckEnvironmentExists("runscope_environment.environmentA"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "name", "test-environment"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "verify_ssl", "true"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.#", "1"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.0.notify_all", "true"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.0.notify_on", "all"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.0.notify_threshold", "1"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.0.recipient.#", "2"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.0.recipient.769210288.name", "bob"),
+					resource.TestCheckResourceAttr("runscope_environment.environmentA", "email.0.recipient.769210288.email", "bob@gmail.com"),
 				),
 			},
 		},
@@ -213,6 +248,50 @@ data "runscope_integration" "slack" {
   type = "slack"
 }
 `
+
+const testRunscopeEnvironmentConfigAWithSingularRemoteAgents = `
+resource "runscope_environment" "environmentA" {
+  bucket_id    = "${runscope_bucket.bucket.id}"
+  name         = "test-environment"
+
+  integrations = [
+		"${data.runscope_integration.slack.id}",
+  ]
+
+  initial_variables = {
+    var1 = "true"
+    var2 = "value2"
+  }
+
+	regions = ["us1", "eu1"]
+
+	remote_agent {
+			name = "test agent"
+			uuid = "arbitrary-string"
+		}
+
+
+	retry_on_failure = true
+	webhooks = ["https://example.com"]
+}
+
+resource "runscope_test" "test" {
+  bucket_id = "${runscope_bucket.bucket.id}"
+  name = "runscope test"
+  description = "This is a test test..."
+}
+
+resource "runscope_bucket" "bucket" {
+  name = "terraform-provider-test"
+  team_uuid = "%s"
+}
+
+data "runscope_integration" "slack" {
+  team_uuid = "%s"
+  type = "slack"
+}
+`
+
 const testRunscopeEnvironmentConfigWithEmail = `
 resource "runscope_environment" "environmentA" {
   bucket_id    = "${runscope_bucket.bucket.id}"
@@ -246,6 +325,63 @@ resource "runscope_environment" "environmentA" {
       		email = "marekpastierik15@gmail.com"
       	}
       	recipients {
+      		name = "bob"
+      		email = "bob@gmail.com"
+      	}
+
+	}
+}
+
+resource "runscope_test" "test" {
+  bucket_id = "${runscope_bucket.bucket.id}"
+  name = "runscope test"
+  description = "This is a test test..."
+}
+
+resource "runscope_bucket" "bucket" {
+  name = "terraform-provider-test"
+  team_uuid = "%s"
+}
+
+data "runscope_integration" "slack" {
+  team_uuid = "%s"
+  type = "slack"
+}
+`
+
+const testRunscopeEnvironmentConfigWithEmailSingular = `
+resource "runscope_environment" "environmentA" {
+  bucket_id    = "${runscope_bucket.bucket.id}"
+  name         = "test-environment"
+
+  integrations = [
+		"${data.runscope_integration.slack.id}"
+  ]
+
+  initial_variables = {
+    var1 = "true"
+    var2 = "value2"
+  }
+
+  regions = ["us1", "eu1"]
+
+	remote_agents {
+			name = "test agent"
+			uuid = "arbitrary-string"
+		}
+
+	retry_on_failure = true
+	webhooks = ["https://example.com"]
+	email {
+	  notify_all       = true
+      notify_on        = "all"
+      notify_threshold = 1
+
+      recipient {
+      		name = "marek"
+      		email = "marekpastierik15@gmail.com"
+      	}
+      	recipient {
       		name = "bob"
       		email = "bob@gmail.com"
       	}